Unify common cases of Array.Copy
